Added animations to boss, and polished fight (really hard)
Boss fight is peepee hard, so we will have to adjust damage, health, size, room, etc

+++ GameDev2/Assets/Scripts/Serializables.cs
@@ -5,7 +5,7 @@ using System;
 
 public class PlayerData
 {
-    public int playerHealth = 5;
+    public float playerHealth = 5;
     public Power mUtil;
     public Power tUtil;
     public Power mWeap;
